Change use of `float` to `CGFloat`.

+++ Butter/BTRSecureTextField.m
@@ -406,7 +406,7 @@ static CGFloat const BTRTextFieldXInset = 2.f;
 		NSSize textSize = [self cellSizeForBounds:theRect];
 		
 		// Center that in the proposed rect
-		float heightDelta = newRect.size.height - textSize.height;
+		CGFloat heightDelta = newRect.size.height - textSize.height;
 		if (heightDelta > 0)
 		{
 			newRect.size.height -= heightDelta;
